[gui] various improvements move fast mode and energy range below plot dont gray out text of item remove unneccessary stretch
